Changed UserActions events so that the current session changing does not emit an event. This fixed a loop on auth pages that listened to UserStore changes, as checkAuth updates the userStore. Also moved login to the nav bar
