+++ src/routes/AuthContainer/components/AuthContainer.js
@@ -19,7 +19,6 @@ export class AuthContainer extends React.Component<void, Props, void> {
     };
     this.onSubmit = this.onSubmit.bind(this)
     this.onChange = this.onChange.bind(this)
-    this.logout = this.logout.bind(this)
     this.checkAuth = this.checkAuth.bind(this)
   }
 
@@ -57,23 +56,6 @@ export class AuthContainer extends React.Component<void, Props, void> {
     });
   }
 
-  logout(e){
-    var user = UserStore.getSessionUser()
-    var scope = this;
-    e.preventDefault();
-    $.ajax({
-      type: "POST",
-      url: "http://test.com:1337/logout",
-      data: {id:user.id, token:user.token},
-      success:function(responseText,requestStatus,fullResponse){
-        if(responseText == "logged out"){
-          UserActions.setSessionUser({id:-1,token:"",status:false})   
-          scope.props.router.push({'pathname':scope.props.redirect})     
-        }
-      }
-    });
-  }
-
   onChange(e){
     var newState = {};
     newState[e.target.id] = e.target.value
@@ -106,18 +88,14 @@ export class AuthContainer extends React.Component<void, Props, void> {
       else{
         display.push(<NoPermission key="noPermission"/>)          
       }
-
-      display.push(<LogoutForm key="logoutForm" logout={this.logout}/>)  
     }
     else{
       var display = (<LoginSignupContainer key="loginOrSignupForm" onChange={this.onChange} onSubmit={this.onSubmit}/> )
     }
 
     return(
-      <div className='container'>
-        <div style={{margin: '0 auto',color:"red"}}>
-          {display}
-        </div>
+      <div >
+        {display}
       </div>
     )
   }
