+++ src/routes/AuthContainer/modules/UserStore.js
@@ -55,7 +55,7 @@ console.log("ryan testing new user store.",action);
 
     case UserConstants.SET_SESSION_USER:
         SESSION_USER = action.user;
-        UserStore.emitChange();
+        //UserStore.emitChange();
         break;
 
     case UserConstants.GET_ALL_USERS:
